Add project documentation and update upload settings
- Added `README.md` containing features, tech stack, architecture overview, and setup instructions.
- Increased `CHUNK_SIZE_IN_MB` from 1 to 10 in `AppConst`.
- Updated `UploadWorkerViewModel` to retrieve progress as a `Float` from `WorkInfo` data instead of `Int`.

+++ app/src/main/java/hrhera/ali/backgroundsync/ui/worker/UploadWorkerViewModel.kt
@@ -57,8 +57,8 @@ class UploadWorkerViewModel @Inject constructor(
                     if (it.state.isFinished) {
                         _progress.value = 100f
                     } else {
-                        val prog = it.progress.getInt(PROGRESS_KEY, 0)
-                        _progress.value = prog.toFloat()
+                        val prog = it.progress.getFloat(PROGRESS_KEY, 0f)
+                        _progress.value = prog
                     }
                 }
             }
@@ -66,3 +66,4 @@ class UploadWorkerViewModel @Inject constructor(
 
 
 }
+
